Глобальная переделка UTCPStream и перевод ModbusTCPXXX на использование writeData/readData (решение проблеммы буферизации). (ModbusTCPServer): первая версия реализации на libev

+++ include/modbus/ModbusRTUSlave.h
@@ -44,7 +44,7 @@ class ModbusRTUSlave:
 	protected:
 
 		// realisation (see ModbusServer.h)
-		virtual int getNextData( unsigned char* buf, int len ) override;
+		virtual size_t getNextData( unsigned char* buf, int len ) override;
 		virtual void setChannelTimeout( timeout_t msec ) override;
 		virtual ModbusRTU::mbErrCode sendData( unsigned char* buf, int len ) override;
